requirements.txt
python-telegram-bot[job-queue]==20.7
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.1.0
psutil==5.9.8
SQLAlchemy==2.0.27
alembic==1.13.1
pandas==2.2.1
matplotlib==3.8.3 